+++ lib/core/widgets/user_devices.dart
@@ -67,13 +67,14 @@ class UserDevices extends HookConsumerWidget {
 
     result.fold(
       (failure) {
+        if (!context.mounted) return;
+        context.hideLoadingDialog();
         context.showSnackBar(failure.localizedErrorMessage);
       },
       (success) async {
         context.showSnackBar('device_removed'.i18n);
-        final innerResult = await ref
-            .read(homeProvider.notifier)
-            .fetchUserData();
+        await ref.read(homeProvider.notifier).fetchUserData();
+        if (!context.mounted) return;
         context.hideLoadingDialog();
       },
     );
